this strange compiler likes operator T* over operator bool() const more and tells you every time a story when he does. But what he likes even more is operator bool() without const. OK, if he wants to ...
svn path=/trunk/koffice/; revision=27667

+++ lib/kscript/kscript_value.cc
@@ -428,7 +428,7 @@ void KSValue::initTypeNameMap()
 {
     if ( typ_to_name ) return;
 
-    typ_to_name = new QString[NTypes];
+    typ_to_name = new QString[(int)NTypes];
 
     typ_to_name[(int)Empty] = QString::fromLatin1("<none>");
     typ_to_name[(int)StringType] = QString::fromLatin1("String");
